neutralize runtime surface and internalize trace formatting
