+++ scripts/eval_rollouts.py
@@ -9,19 +9,21 @@ from pathlib import Path
 from tempfile import TemporaryDirectory
 from typing import Any
 
+from open_range._reference_replay import (
+    action_for_reference_step,
+    reference_trace_pairs,
+)
 from open_range._runtime_store import hydrate_runtime_snapshot
 from open_range.build_config import OFFLINE_BUILD_CONFIG
 from open_range.curriculum import FrontierMutationPolicy, PopulationStats
 from open_range.episode_config import EpisodeConfig
 from open_range.pipeline import BuildPipeline
-from open_range.probe_planner import runtime_action
 from open_range.resources import load_bundled_manifest
-from open_range.runtime import ReferenceDrivenRuntime
-from open_range.runtime_types import Action, EpisodeScore
+from open_range.runtime import OpenRangeRuntime
+from open_range.runtime_types import EpisodeScore
 from open_range.sim import ReferenceSimPlane
 from open_range.snapshot import RuntimeSnapshot
 from open_range.store import FileSnapshotStore
-from open_range.training_data import normalize_trace_action
 
 
 def _default_manifest_name() -> str:
@@ -46,10 +48,8 @@ def _run_mode(
     snapshot: RuntimeSnapshot, episode_config: EpisodeConfig
 ) -> dict[str, Any]:
     pair_reports: list[dict[str, Any]] = []
-    for attack_idx, defense_idx in _reference_trace_pairs(
-        snapshot, episode_config.mode
-    ):
-        runtime = ReferenceDrivenRuntime()
+    for attack_idx, defense_idx in reference_trace_pairs(snapshot, episode_config.mode):
+        runtime = OpenRangeRuntime()
         runtime.reset(
             snapshot,
             episode_config,
@@ -67,7 +67,7 @@ def _run_mode(
                 raise
             runtime.act(
                 decision.actor,
-                _reference_action(
+                action_for_reference_step(
                     snapshot,
                     decision.actor,
                     runtime.reference_step(decision.actor),
@@ -197,25 +197,6 @@ def _population_stats(
         novelty=novelty,
         blue_signal_points=snapshot.validator_report.blue_signal_points,
     )
-
-
-def _reference_action(snapshot: RuntimeSnapshot, actor: str, expected) -> Action:
-    if expected is None:
-        return Action(actor_id=actor, role=actor, kind="sleep", payload={})
-    return normalize_trace_action(snapshot, runtime_action(actor, expected))
-
-
-def _reference_trace_pairs(
-    snapshot: RuntimeSnapshot, mode: str
-) -> tuple[tuple[int, int], ...]:
-    attack_count = max(1, len(snapshot.reference_bundle.reference_attack_traces))
-    defense_count = max(1, len(snapshot.reference_bundle.reference_defense_traces))
-    if mode == "red_only":
-        return tuple((idx, idx % defense_count) for idx in range(attack_count))
-    if mode in {"blue_only_live", "blue_only_from_prefix"}:
-        return tuple((idx % attack_count, idx) for idx in range(defense_count))
-    count = max(attack_count, defense_count)
-    return tuple((idx % attack_count, idx % defense_count) for idx in range(count))
 
 
 def evaluate_rollouts(
