+++ src/open_range/service.py
@@ -1,4 +1,4 @@
-"""Thin store-backed runtime facade for admitted snapshots."""
+"""Thin store-backed decision-loop facade for admitted snapshots."""
 
 from __future__ import annotations
 
@@ -8,7 +8,7 @@ from open_range._runtime_store import load_runtime_snapshot, sample_runtime_snap
 from open_range.cluster import BootedRelease, LiveBackend
 from open_range.episode_config import DEFAULT_EPISODE_CONFIG, EpisodeConfig
 from open_range.execution import ActionBackend, PodActionBackend
-from open_range.runtime import ReferenceDrivenRuntime
+from open_range.runtime import OpenRangeRuntime
 from open_range.runtime_types import (
     Action,
     ActionResult,
@@ -26,12 +26,12 @@ class OpenRange:
         self,
         *,
         store: FileSnapshotStore | None = None,
-        runtime: ReferenceDrivenRuntime | None = None,
+        runtime: OpenRangeRuntime | None = None,
         live_backend: LiveBackend | None = None,
         action_backend: ActionBackend | None = None,
     ) -> None:
         self.store = store or FileSnapshotStore()
-        self.runtime = runtime or ReferenceDrivenRuntime()
+        self.runtime = runtime or OpenRangeRuntime()
         self.live_backend = live_backend
         self.action_backend = action_backend or PodActionBackend()
         self._active_snapshot_id = ""
